#50: backend create business logic 50: backend create business logic 50: backend create business logic

+++ backend-core/src/main/resources/db/migration/V1.2.0.35__model_media_viewing_feature.sql
@@ -83,7 +83,7 @@ CREATE TABLE IF NOT EXISTS media_rating
     username         VARCHAR(100)                NOT NULL,
     media_id         SERIAL                      NOT NULL,
     media_flavour_id VARCHAR(5)                  NOT NULL,
-    score            SMALLINT,
+    score            FLOAT,
     created_at       TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now()::TIMESTAMP,
     modified_at      TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now()::TIMESTAMP
 );
